FEBio2: synced with FEBio (added multiphasic code)
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@3462 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioLib/FESolute.h
@@ -118,8 +118,8 @@ public:
 	//! solute molecular weight
 	double MolarMass() { return m_M; }
 	
-	//! solute valence
-	double ChargeNumber() { return m_z; }
+	//! solute charge number
+	int ChargeNumber() { return m_z; }
 	
 	//! Serialization
 	void Serialize(DumpFile& ar);
